# frontend/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AddWorkspaceRootRequest": {
        "additionalProperties": false,
        "description": "One user-authorized additional workspace directory.",
        "properties": {
          "path": {
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "AddWorkspaceRootRequest",
        "type": "object"
      },
...
      "ApiErrorCode": {
        "description": "Stable HTTP API errors not emitted by a graph run.",
        "enum": [
          "invalid_request",
          "internal_error",
          "handoff_not_pending",
          "tool_approval_not_pending",
          "session_already_exists",
          "session_busy",
          "session_not_found",
          "knowledge_unavailable"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
...
      "DailyAccuracyPoint": {
        "additionalProperties": false,
        "description": "学情洞察：一个 UTC 日的作答量与加权正确率（赛前可视化增强）。\n\n加权口径与 DiagnosisSummary 同源（correct×1 + partial×0.5，见\ncore/learning/store.py insights），确定性聚合、可复现。",
        "properties": {
          "accuracy": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Accuracy",
            "type": "number"
          },
          "attempts": {
            "minimum": 1.0,
            "title": "Attempts",
            "type": "integer"
          },
          "date": {
            "minLength": 1,
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "attempts",
          "accuracy"
        ],
        "title": "DailyAccuracyPoint",
        "type": "object"
      },
      "DiagnosisKnowledgePoint": {
        "additionalProperties": false,
        "description": "学情诊断：一个知识点的作答聚合（六大功能 P3-15）。",
        "properties": {
          "accuracy": {
            "maximum": 1,
            "minimum": 0,
            "title": "Accuracy",
            "type": "number"
          },
          "attempts": {
            "minimum": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "correct": {
            "minimum": 0,
            "title": "Correct",
            "type": "integer"
          },
          "knowledge_point": {
            "title": "Knowledge Point",
            "type": "string"
          },
          "last_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Last At"
          }
        },
        "required": [
          "knowledge_point",
          "attempts",
          "correct",
          "accuracy"
        ],
        "title": "DiagnosisKnowledgePoint",
        "type": "object"
      },
      "DiagnosisSummary": {
        "additionalProperties": false,
        "description": "学情诊断摘要（六大功能 P3-15）。\n\n数据源是 learning_records 的 SQL 聚合（确定性规则：预警 =\nattempts≥2 且加权正确率<0.6，见 core/learning/store.py），LLM\n叙述只出现在对话内诊断报告，不在本契约里。",
        "properties": {
          "knowledge_points": {
            "items": {
              "$ref": "#/components/schemas/DiagnosisKnowledgePoint"
            },
            "title": "Knowledge Points",
            "type": "array"
          },
          "total_attempts": {
            "default": 0,
            "minimum": 0,
            "title": "Total Attempts",
            "type": "integer"
          },
          "uncategorized_attempts": {
            "default": 0,
            "minimum": 0,
            "title": "Uncategorized Attempts",
            "type": "integer"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "User Id"
          },
          "weak_points": {
            "items": {
              "type": "string"
            },
            "title": "Weak Points",
            "type": "array"
          }
        },
        "title": "DiagnosisSummary",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "A sanitized stable API error.",
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "FastAPI's standard error envelope.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningInsights": {
        "additionalProperties": false,
        "description": "学情洞察摘要（学习进度页错题/趋势/路径卡的数据源）。\n\n与 DiagnosisSummary 的分工：后者面向「知识点掌握与预警」（诊断），\n本契约面向「错题归因分布 / 正确率趋势 / 路径存档回显」（展示）；\n数据源同为 learning_records 的确定性 SQL 聚合（有界窗口：趋势近 30 日、\n路径近 20 条）。store 未注入/无记录时返回空报告 200（降级红线同诊断端点）。",
        "properties": {
          "daily_accuracy": {
            "items": {
              "$ref": "#/components/schemas/DailyAccuracyPoint"
            },
            "title": "Daily Accuracy",
            "type": "array"
          },
          "error_tag_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Error Tag Counts",
            "type": "object"
          },
          "recent_path_plans": {
            "items": {
              "$ref": "#/components/schemas/PathPlanRecordDto"
            },
            "title": "Recent Path Plans",
            "type": "array"
          },
          "total_wrong": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Wrong",
            "type": "integer"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "title": "LearningInsights",
        "type": "object"
      },
...
      "PathPlanRecordDto": {
        "additionalProperties": false,
        "description": "学习路径存档回显项（只有知识点与时间，无路径正文——脱敏口径）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "knowledge_point": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Point"
          }
        },
        "title": "PathPlanRecordDto",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceDirectory": {
        "additionalProperties": false,
        "description": "One directory entry in the server-side workspace picker.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path"
        ],
        "title": "WorkspaceDirectory",
        "type": "object"
      },
      "WorkspaceDirectoryListing": {
        "additionalProperties": false,
        "description": "Directory picker state rooted in the server filesystem.",
        "properties": {
          "directories": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceDirectory"
            },
            "title": "Directories",
            "type": "array"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "WorkspaceDirectoryListing",
        "type": "object"
      },
      "WorkspacePath": {
        "additionalProperties": false,
        "description": "A canonical existing directory accepted by the server policy.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name"
        ],
        "title": "WorkspacePath",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/learning/diagnosis/summary": {
      "get": {
        "description": "返回学情诊断摘要（默认当前用户；student_id 供教师视角查询）。\n\n边界与降级见模块 docstring：v1 的 student_id 查询是演示环境可信\n声明（无鉴权），生产需待认证落地；store 缺失或空数据返回空报告。",
        "operationId": "diagnosis_summary_learning_diagnosis_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnosis Summary",
        "tags": [
          "learning"
        ]
      }
    },
    "/learning/insights/summary": {
      "get": {
        "description": "返回学情洞察摘要（错题归因/正确率趋势/路径存档回显）。\n\n降级红线与诊断端点一致：store 缺失或空数据返回空报告 200。\n数据窗口有界（趋势近 30 日、路径近 20 条，见 core/learning/store.py）。",
        "operationId": "learning_insights_learning_insights_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningInsights"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Learning Insights",
        "tags": [
          "learning"
        ]
      }
    },
...
    "/workspaces/directories": {
      "get": {
        "description": "Browse immediate directories without exposing files or file contents.",
        "operationId": "list_workspace_directories_workspaces_directories_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Path"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDirectoryListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Workspace Directories",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/workspaces/validate": {
      "post": {
        "description": "Resolve a typed path before the user creates or updates a session.",
        "operationId": "validate_workspace_workspaces_validate_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWorkspaceRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePath"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Validate Workspace",
        "tags": [
          "workspaces"
        ]
      }
    }
  },
...
}